refactor(sdk): move ContractKind into types, drop unused re-exports, tidy tests
