+++ sdk/native/src/chain/mod.rs
@@ -10,7 +10,6 @@ mod submit;
 mod tx_assemble;
 mod tx_prepare;
 
-pub use contract_error::{ContractErrorInfo, ContractKind, parse_contract_error, translate};
 pub use contract_state::{PreparedSorobanTx, StateFetcher};
 pub use indexer::ContractDataStorage;
 pub use rpc::{Client, Client as RpcClient};
